Remove old code and render individual categories
