+++ src/main.rs
@@ -1,126 +1,27 @@
-mod cache;
-mod caching;
-mod category;
-mod headers;
-mod i18n;
-mod redirect;
-mod rust_version;
-mod teams;
-mod render;
-
-use cache::Cache;
-use cache::Cached;
-use rocket::catch;
-use rocket::get;
-use rocket::tokio::sync::RwLock;
-use rust_version::RustVersion;
+use crate::assets::{AssetFiles, compile_assets};
+use crate::category::Category;
+use crate::fs::ensure_directory;
+use crate::i18n::{EXPLICIT_LOCALE_INFO, LocaleInfo, SUPPORTED_LOCALES, TeamHelper, create_loader};
+use crate::rust_version::{RustVersion, fetch_rust_version};
+use crate::teams::{RustTeams, load_rust_teams};
+use anyhow::Context;
+use handlebars::{DirectorySourceOptions, Handlebars};
+use handlebars_fluent::{FluentHelper, Loader, SimpleLoader};
 use serde::Serialize;
-use teams::RustTeams;
-use teams::encode_zulip_stream;
-
-use std::collections::hash_map::DefaultHasher;
-use std::env;
-use std::fs;
 use std::fs::File;
-use std::hash::Hasher;
 use std::io::BufWriter;
 use std::path::{Path, PathBuf};
-use std::sync::Arc;
-use std::sync::LazyLock;
-use anyhow::Context as _;
-use handlebars::{DirectorySourceOptions, Handlebars};
-use rocket::{
-    fs::NamedFile,
-    http::Status,
-    request::{FromParam, Request},
-    response::{Redirect, content},
-};
-use rocket_dyn_templates::Template;
 
-use sass_rs::{Options, compile_file};
+mod assets;
+mod category;
+mod fs;
+mod i18n;
+mod rust_version;
+mod teams;
 
-use category::Category;
-
-use caching::CachedNamedFile;
-use handlebars_fluent::{FluentHelper, loader::Loader, simple_loader};
-use i18n::{EXPLICIT_LOCALE_INFO, LocaleInfo, SupportedLocale, TeamHelper, create_loader};
+static PONTOON_ENABLED: bool = false;
 
 const ZULIP_DOMAIN: &str = "https://rust-lang.zulipchat.com";
-
-static ASSETS: LazyLock<AssetFiles> = LazyLock::new(|| {
-    let app_css_file = compile_sass("app");
-    let fonts_css_file = compile_sass("fonts");
-    let vendor_css_file = concat_vendor_css(vec!["tachyons"]);
-    let app_js_file = concat_app_js(vec!["tools-install"]);
-
-    AssetFiles {
-        css: CSSFiles {
-            app: app_css_file,
-            fonts: fonts_css_file,
-            vendor: vendor_css_file,
-        },
-        js: JSFiles { app: app_js_file },
-    }
-});
-static PONTOON_ENABLED: LazyLock<bool> = LazyLock::new(|| env::var("RUST_WWW_PONTOON").is_ok());
-static ROBOTS_TXT_DISALLOW_ALL: LazyLock<bool> =
-    LazyLock::new(|| env::var("ROBOTS_TXT_DISALLOW_ALL").is_ok());
-
-#[derive(Serialize)]
-struct Context<T: Serialize> {
-    page: String,
-    title: String,
-    parent: &'static str,
-    is_landing: bool,
-    data: T,
-    lang: String,
-    baseurl: String,
-    pontoon_enabled: bool,
-    assets: &'static AssetFiles,
-    locales: &'static [LocaleInfo],
-    is_translation: bool,
-}
-
-impl<T: Serialize> Context<T> {
-    fn new(page: &str, title_id: &str, is_landing: bool, data: T, lang: String) -> Self {
-        let helper = create_loader();
-        let title = if title_id.is_empty() {
-            "".into()
-        } else {
-            let lang = lang.parse().expect("lang should be valid");
-            helper.lookup(&lang, title_id, None)
-        };
-        Self {
-            page: page.to_owned(),
-            title,
-            parent: LAYOUT,
-            is_landing,
-            data,
-            baseurl: baseurl(&lang),
-            is_translation: lang != "en-US",
-            lang,
-            pontoon_enabled: *PONTOON_ENABLED,
-            assets: &ASSETS,
-            locales: EXPLICIT_LOCALE_INFO,
-        }
-    }
-}
-
-#[derive(Clone, Serialize)]
-struct CSSFiles {
-    app: String,
-    fonts: String,
-    vendor: String,
-}
-#[derive(Clone, Serialize)]
-struct JSFiles {
-    app: String,
-}
-#[derive(Clone, Serialize)]
-struct AssetFiles {
-    css: CSSFiles,
-    js: JSFiles,
-}
 
 static LAYOUT: &str = "components/layout";
 static ENGLISH: &str = "en-US";
@@ -133,334 +34,227 @@ fn baseurl(lang: &str) -> String {
     }
 }
 
-#[get("/logos/<file..>", rank = 1)]
-async fn logos(file: PathBuf) -> Option<CachedNamedFile> {
-    NamedFile::open(Path::new("static/logos").join(file))
-        .await
-        .ok()
-        .map(|file| CachedNamedFile::max_age(file, 3600))
+#[derive(Serialize)]
+struct TemplateCtx<'a, T: Serialize> {
+    page: String,
+    title: String,
+    parent: &'static str,
+    is_landing: bool,
+    data: &'a T,
+    lang: String,
+    baseurl: String,
+    pontoon_enabled: bool,
+    assets: &'a AssetFiles,
+    locales: &'static [LocaleInfo],
+    is_translation: bool,
 }
 
-#[get("/static/<file..>", rank = 1)]
-async fn files(file: PathBuf) -> Option<CachedNamedFile> {
-    NamedFile::open(Path::new("static/").join(file))
-        .await
-        .ok()
-        .map(|file| CachedNamedFile::max_age(file, 3600))
+struct PageCtx<'a, T: Serialize> {
+    template_ctx: TemplateCtx<'a, T>,
+    output_dir: &'a Path,
+    handlebars: &'a Handlebars<'a>,
 }
 
-#[get("/robots.txt", rank = 1)]
-fn robots_txt() -> Option<content::RawText<&'static str>> {
-    if *ROBOTS_TXT_DISALLOW_ALL {
-        Some(content::RawText("User-agent: *\nDisallow: /"))
-    } else {
-        None
+impl<'a, T: Serialize> PageCtx<'a, T> {
+    fn make_landing(mut self) -> Self {
+        self.template_ctx.is_landing = true;
+        self
+    }
+
+    fn render<P: AsRef<Path>>(self, dst_path: P) -> anyhow::Result<()> {
+        let path = dst_path.as_ref();
+        let template = &self.template_ctx.page;
+
+        let out_path = self.output_dir.join(path);
+        ensure_directory(&out_path)?;
+        let mut output_file = BufWriter::new(File::create(&out_path)?);
+        eprintln!("Rendering `{template}` into {}", out_path.display());
+
+        self.handlebars
+            .render_to_write(template, &self.template_ctx, &mut output_file)
+            .with_context(|| {
+                anyhow::anyhow!(
+                    "cannot render template {template} into {}",
+                    out_path.display()
+                )
+            })?;
+        Ok(())
     }
 }
 
-#[get("/")]
-async fn index(version_cache: &Cache<RustVersion>) -> Template {
-    render_index(ENGLISH.into(), version_cache).await
+struct RenderCtx<'a> {
+    handlebars: Handlebars<'a>,
+    fluent_loader: SimpleLoader,
+    output_dir: PathBuf,
+    rust_version: RustVersion,
+    teams: RustTeams,
+    assets: AssetFiles,
 }
 
-#[get("/<locale>", rank = 3)]
-async fn index_locale(locale: SupportedLocale, version_cache: &Cache<RustVersion>) -> Template {
-    render_index(locale.0, version_cache).await
-}
-
-#[get("/<category>")]
-fn category_en(category: Category) -> Template {
-    render_category(category, ENGLISH.into())
-}
-
-#[get("/<locale>/<category>", rank = 9)]
-fn category_locale(category: Category, locale: SupportedLocale) -> Template {
-    render_category(category, locale.0)
-}
-
-#[get("/governance")]
-async fn governance(teams_cache: &Cache<RustTeams>) -> Result<Template, Status> {
-    render_governance(ENGLISH.into(), teams_cache).await
-}
-
-#[get("/governance/<section>/<team>", rank = 2)]
-async fn team(
-    section: &str,
-    team: &str,
-    teams_cache: &Cache<RustTeams>,
-) -> Result<Template, Status> {
-    render_team(section, team, ENGLISH.into(), teams_cache).await
-}
-
-#[get("/<locale>/governance", rank = 8)]
-async fn governance_locale(
-    locale: SupportedLocale,
-    teams_cache: &Cache<RustTeams>,
-) -> Result<Template, Status> {
-    render_governance(locale.0, teams_cache).await
-}
-
-#[get("/<locale>/governance/<section>/<team>", rank = 12)]
-async fn team_locale(
-    section: &str,
-    team: &str,
-    locale: SupportedLocale,
-    teams_cache: &Cache<RustTeams>,
-) -> Result<Template, Status> {
-    render_team(section, team, locale.0, teams_cache).await
-}
-
-#[get("/<category>/<subject>", rank = 4)]
-fn subject(category: Category, subject: &str) -> Result<Template, Status> {
-    render_subject(category, subject, ENGLISH.into())
-}
-
-#[get("/<locale>/<category>/<subject>", rank = 14)]
-fn subject_locale(
-    category: Category,
-    subject: &str,
-    locale: SupportedLocale,
-) -> Result<Template, Status> {
-    render_subject(category, subject, locale.0)
-}
-
-#[get("/en-US", rank = 1)]
-fn redirect_bare_en_us() -> Redirect {
-    Redirect::permanent("/")
-}
-
-#[get("/.well-known/security.txt")]
-fn well_known_security() -> &'static str {
-    include_str!("../static/text/well_known_security.txt")
-}
-
-#[catch(404)]
-#[allow(clippy::result_large_err)]
-fn not_found(req: &Request) -> Result<Template, Redirect> {
-    if let Some(redirect) = crate::redirect::maybe_redirect(req.uri().path()) {
-        return Err(redirect);
-    }
-
-    let lang = if let Some(next) = req.uri().path().segments().next() {
-        if let Ok(lang) = SupportedLocale::from_param(next) {
-            lang.0
+impl<'a> RenderCtx<'a> {
+    fn page<T: Serialize>(
+        &'a self,
+        page: &str,
+        title_id: &str,
+        data: &'a T,
+        lang: &str,
+    ) -> PageCtx<'a, T> {
+        let title = if title_id.is_empty() {
+            "".into()
         } else {
-            ENGLISH.into()
+            let lang = lang.parse().expect("lang should be valid");
+            self.fluent_loader.lookup(&lang, title_id, None)
+        };
+        PageCtx {
+            template_ctx: TemplateCtx {
+                page: page.to_string(),
+                title,
+                parent: LAYOUT,
+                is_landing: false,
+                data,
+                baseurl: baseurl(&lang),
+                is_translation: lang != "en-US",
+                lang: lang.to_string(),
+                pontoon_enabled: PONTOON_ENABLED,
+                assets: &self.assets,
+                locales: EXPLICIT_LOCALE_INFO,
+            },
+            output_dir: &self.output_dir,
+            handlebars: &self.handlebars,
         }
-    } else {
-        ENGLISH.into()
+    }
+
+    fn copy_static_assets<P: AsRef<Path>>(&self, src_dir: P, dst_dir: P) -> anyhow::Result<()> {
+        let dst = self.output_dir.join(dst_dir.as_ref());
+        println!(
+            "Copying static assets from {} to {}",
+            src_dir.as_ref().display(),
+            dst.display()
+        );
+        copy_dir_all(src_dir.as_ref(), dst)?;
+        Ok(())
+    }
+}
+
+/// Calls `func` for all supported languages.
+/// Passes it the destination path into which should a given page be rendered, and the language
+/// in which it should be rendered.
+fn all_langs<F>(dst_path: &str, func: F) -> anyhow::Result<()>
+where
+    F: Fn(&str, &str) -> anyhow::Result<()>,
+{
+    for lang in SUPPORTED_LOCALES.iter() {
+        let path = match lang {
+            l if *l == ENGLISH => dst_path.to_string(),
+            l => format!("{l}/{dst_path}"),
+        };
+        func(&path, lang).with_context(|| anyhow::anyhow!("could not handle language {lang}"))?;
+    }
+    Ok(())
+}
+
+fn copy_dir_all(src: impl AsRef<Path>, dst: impl AsRef<Path>) -> std::io::Result<()> {
+    std::fs::create_dir_all(&dst)?;
+    for entry in std::fs::read_dir(src)? {
+        let entry = entry?;
+        let ty = entry.file_type()?;
+        if ty.is_dir() {
+            copy_dir_all(entry.path(), dst.as_ref().join(entry.file_name()))?;
+        } else {
+            std::fs::copy(entry.path(), dst.as_ref().join(entry.file_name()))?;
+        }
+    }
+    Ok(())
+}
+
+fn setup_handlebars() -> anyhow::Result<Handlebars<'static>> {
+    let mut handlebars: Handlebars<'static> = Handlebars::new();
+    handlebars.set_strict_mode(true);
+
+    let mut options = DirectorySourceOptions::default();
+    options.tpl_extension = ".html.hbs".to_string();
+    handlebars
+        .register_templates_directory("templates", options)
+        .context("cannot register template directory")?;
+
+    let loader = create_loader();
+    let helper = FluentHelper::new(loader);
+    handlebars.register_helper("fluent", Box::new(helper));
+    handlebars.register_helper("team-text", Box::new(TeamHelper::new()));
+    Ok(handlebars)
+}
+
+#[tokio::main]
+async fn main() -> anyhow::Result<()> {
+    let rust_version = fetch_rust_version().await?;
+    let teams = load_rust_teams().await?;
+
+    // Prepare build directory
+    let output_dir = PathBuf::from("html");
+    let _ = std::fs::remove_dir_all(&output_dir);
+    std::fs::create_dir_all(&output_dir)?;
+
+    let assets = compile_assets(Path::new("."), &output_dir, "/")?;
+    let handlebars = setup_handlebars()?;
+
+    let ctx = RenderCtx {
+        fluent_loader: create_loader(),
+        assets,
+        rust_version,
+        teams,
+        handlebars,
+        output_dir,
     };
+    ctx.copy_static_assets("static", "static")?;
 
-    Ok(not_found_locale(lang))
+    render_index(&ctx)?;
+    render_governance(&ctx)?;
+    render_category(&ctx, "community")?;
+    render_category(&ctx, "learn")?;
+    render_category(&ctx, "policies")?;
+    render_category(&ctx, "tools")?;
+    render_category(&ctx, "what")?;
+
+    // TODO: 404, redirects
+
+    Ok(())
 }
 
-#[catch(422)]
-#[allow(clippy::result_large_err)]
-fn unprocessable_content(req: &Request) -> Result<Template, Redirect> {
-    not_found(req)
-}
-
-fn not_found_locale(lang: String) -> Template {
-    let page = "404";
-    let context = Context::new(page, "error404-page-title", false, (), lang);
-    Template::render(page, context)
-}
-
-#[catch(500)]
-fn catch_error() -> Template {
-    not_found_locale(ENGLISH.into())
-}
-
-fn hash_css(css: &str) -> String {
-    let mut hasher = DefaultHasher::new();
-    hasher.write(css.as_bytes());
-    hasher.finish().to_string()
-}
-
-fn compile_sass(filename: &str) -> String {
-    let scss_file = format!("./src/styles/{filename}.scss");
-
-    let css = compile_file(&scss_file, Options::default())
-        .unwrap_or_else(|_| panic!("couldn't compile sass: {}", &scss_file));
-
-    let css_sha = format!("{}_{}", filename, hash_css(&css));
-    let css_file = format!("./static/styles/{css_sha}.css");
-
-    fs::write(&css_file, css.into_bytes())
-        .unwrap_or_else(|_| panic!("couldn't write css file: {}", &css_file));
-
-    String::from(&css_file[1..])
-}
-
-fn concat_vendor_css(files: Vec<&str>) -> String {
-    let mut concatted = String::new();
-    for filestem in files {
-        let vendor_path = format!("./static/styles/{filestem}.css");
-        let contents = fs::read_to_string(vendor_path).expect("couldn't read vendor css");
-        concatted.push_str(&contents);
-    }
-
-    let css_sha = format!("vendor_{}", hash_css(&concatted));
-    let css_path = format!("./static/styles/{}.css", &css_sha);
-
-    fs::write(&css_path, &concatted).expect("couldn't write vendor css");
-
-    String::from(&css_path[1..])
-}
-
-fn concat_app_js(files: Vec<&str>) -> String {
-    let mut concatted = String::new();
-    for filestem in files {
-        let vendor_path = format!("./static/scripts/{filestem}.js");
-        let contents = fs::read_to_string(vendor_path).expect("couldn't read app js");
-        concatted.push_str(&contents);
-    }
-
-    let js_sha = format!("app_{}", hash_css(&concatted));
-    let js_path = format!("./static/scripts/{}.js", &js_sha);
-
-    fs::write(&js_path, &concatted).expect("couldn't write app js");
-
-    String::from(&js_path[1..])
-}
-
-async fn render_index(lang: String, version_cache: &Cache<RustVersion>) -> Template {
+fn render_index(render_ctx: &RenderCtx) -> anyhow::Result<()> {
     #[derive(Serialize)]
     struct IndexData {
         rust_version: String,
     }
-
-    let page = "index";
     let data = IndexData {
-        rust_version: rust_version::rust_version(version_cache).await,
+        rust_version: render_ctx.rust_version.0.clone(),
     };
-    let context = Context::new(page, "", true, data, lang);
-    Template::render(page, context)
+    all_langs("index.html", |path, lang| {
+        render_ctx
+            .page("index", "", &data, lang)
+            .make_landing()
+            .render(path)
+    })
 }
 
-fn render_category(category: Category, lang: String) -> Template {
-    let page = category.index();
-    let title_id = format!("{}-page-title", category.name());
-    let context = Context::new(category.name(), &title_id, false, (), lang);
+fn render_governance(render_ctx: &RenderCtx) -> anyhow::Result<()> {
+    let data = render_ctx.teams.index_data();
 
-    Template::render(page, context)
+    all_langs("governance/index.html", |dst_path, lang| {
+        render_ctx
+            .page("governance/index", "governance-page-title", &data, lang)
+            .render(dst_path)
+    })
 }
 
-async fn render_governance(
-    lang: String,
-    teams_cache: &Cache<RustTeams>,
-) -> Result<Template, Status> {
-    match teams::index_data(teams_cache).await {
-        Ok(data) => {
-            let page = "governance/index";
-            let context = Context::new(page, "governance-page-title", false, data, lang);
-
-            Ok(Template::render(page, context))
-        }
-        Err(err) => {
-            eprintln!("error while loading the governance page: {err}");
-            Err(Status::InternalServerError)
-        }
-    }
-}
-
-async fn render_team(
-    section: &str,
-    team: &str,
-    lang: String,
-    teams_cache: &Cache<RustTeams>,
-) -> Result<Template, Status> {
-    match teams::page_data(section, team, teams_cache).await {
-        Ok(data) => {
-            let page = "governance/group";
-            let name = format!("governance-team-{}-name", data.team.name);
-            let context = Context::new(page, &name, false, data, lang);
-            Ok(Template::render(page, context))
-        }
-        Err(err) => {
-            if err.is::<teams::TeamNotFound>() {
-                Err(Status::NotFound)
-            } else {
-                eprintln!("error while loading the team page: {err}");
-                Err(Status::InternalServerError)
-            }
-        }
-    }
-}
-
-fn render_subject(category: Category, subject: &str, lang: String) -> Result<Template, Status> {
-    // Rocket's Template::render method is not really designed to accept arbitrary templates: if a
-    // template is missing, it just returns a Status::InternalServerError, without a way to
-    // distinguish it from a syntax error in the template itself.
-    //
-    // To work around the problem we check whether the template exists beforehand.
-    let path = Path::new("templates")
-        .join(category.name())
-        .join(format!("{subject}.html.hbs"));
-    if !path.is_file() {
-        return Err(Status::NotFound);
-    }
-
-    let page = format!("{}/{}", category.name(), subject);
-    let title_id = format!("{}-{}-page-title", category.name(), subject);
-    let context = Context::new(subject, &title_id, false, (), lang);
-
-    Ok(Template::render(page, context))
-}
-
-// #[rocket::launch]
-// async fn rocket() -> _ {
-//     let templating = Template::custom(|engine| {
-//         engine
-//             .handlebars
-//             .register_helper("fluent", Box::new(FluentHelper::new(create_loader())));
-//         engine
-//             .handlebars
-//             .register_helper("team-text", Box::new(TeamHelper::new()));
-//         engine
-//             .handlebars
-//             .register_helper("encode-zulip-stream", Box::new(encode_zulip_stream));
-//     });
-//
-//     let rust_version = RustVersion::fetch().await.unwrap_or_default();
-//     let teams = RustTeams::fetch().await.unwrap_or_default();
-//
-//     rocket::build()
-//         .attach(templating)
-//         .attach(headers::InjectHeaders)
-//         .manage(Arc::new(RwLock::new(rust_version)))
-//         .manage(Arc::new(RwLock::new(teams)))
-//         .mount(
-//             "/",
-//             rocket::routes![
-//                 index,
-//                 category_en,
-//                 governance,
-//                 team,
-//                 subject,
-//                 files,
-//                 robots_txt,
-//                 logos,
-//                 index_locale,
-//                 category_locale,
-//                 governance_locale,
-//                 team_locale,
-//                 subject_locale,
-//                 redirect_bare_en_us,
-//                 well_known_security,
-//             ],
-//         )
-//         .register(
-//             "/",
-//             rocket::catchers![not_found, unprocessable_content, catch_error],
-//         )
-// }
-
-#[tokio::main]
-async fn main() -> anyhow::Result<()> {
-    let rust_version = RustVersion::fetch().await?;
-    let teams = RustTeams::fetch().await?;
-    render::render(rust_version, teams)?;
-
-    Ok(())
+fn render_category(render_ctx: &RenderCtx, category: &str) -> anyhow::Result<()> {
+    all_langs(&format!("{category}/index.html"), |dst_path, lang| {
+        render_ctx
+            .page(
+                &format!("{category}/index"),
+                &format!("{category}-page-title"),
+                &(),
+                lang,
+            )
+            .render(dst_path)
+    })
 }
